% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{fPlot}
\alias{fPlot}
\title{nice Seurat feature plot}
\usage{
fPlot(object, par, width = 16, height = ceiling(length(genes)/2))
}
\arguments{
\item{object}{Seurat object}

\item{par}{column name in markers.csv}

\item{width}{width of output plot (default: 16)}

\item{height}{height of output plot (default: length of genes divided by two)}
}
\value{
save feature plot to folder \code{featureplot}
}
\description{
create and save a nice Seurat feature plot in folder \code{featureplot}
}
\examples{
\dontrun{fPlot(sc_merge, par = "main", width = 5, height = 5)}
}
