% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{fPlot}
\alias{fPlot}
\title{nice Seurat feature plot}
\usage{
fPlot(object, par, width = 16, height = ceiling(length(genes_found)/4) * 3)
}
\arguments{
\item{object}{Seurat object}

\item{par}{column name in markers.csv}

\item{width}{width of output plot (default: 16)}

\item{height}{height of output plot (default: length of genes divided by four, ceiling, times three)}
}
\value{
save feature plot to folder \verb{/results/featureplot/}
}
\description{
create and save a nice Seurat feature plot in folder \code{featureplot}
}
\examples{
\dontrun{fPlot(sc_merge, par = "main", filepath = file.path("results", "featureplot", glue::glue("fp_")))}
}
